borrar todo lo innecesario
